refactor: test 패키지 구조 main과 동일하게 리팩터
